+++ HISTORY.md
@@ -5,6 +5,7 @@ https://github.com/enmasseio/babble
 ## not yet released, version 0.10.0
 
 - Implemented function `Babbler.listenOnce`.
+- Renamed default message listener for babblify from `onMessage` to `receive`.
 
 
 ## 2014-08-11, version 0.9.0